refactor: rename files for queues to be more concise
