+++ int/redis/tests/test_queue.py
@@ -7,7 +7,7 @@ import json
 from redis.asyncio import Redis
 import time
 
-from acapy_plugin_pickup.undelivered_queue.redis_persisted_queue import (
+from acapy_plugin_pickup.undelivered_queue.redis import (
     RedisUndeliveredQueue,
 )
 from acapy_plugin_pickup.undelivered_queue.base import message_id_for_outbound
